app/v1/utils/settings.py
import os

from pydantic import BaseSettings
from dotenv import load_dotenv

load_dotenv()


class Settings(BaseSettings):
    db_name: str = os.getenv("DB_NAME", "")
    db_user: str = os.getenv("DB_USER", "")
    db_password: str = os.getenv("DB_PASS", "")
    db_host: str = os.getenv("DB_HOST", "")
    db_port: str = os.getenv("DB_PORT", "")

    def __init__(self, **kwargs):
        super().__init__(**kwargs)
        if self.db_name is None:
            raise ValueError("DB_NAME is not set")
