requirements.txt
fastapi
uvicorn
SQLAlchemy
requests
psycopg2
python-dotenv